pyproject.toml
[project]
name = "caddy"
version = "2.4.0"
description = "Lightweight application service to harvest and index the cadastre dataset, and expose an API endpoint to allow full-text searching of addresses."
authors = [{ name = "Ashley Felton", email = "ashley.felton@dbca.wa.gov.au" }]
license = "Apache-2.0"
requires-python = ">=3.13,<4.0"
dependencies = [
  "django==4.2.21",
  "psycopg[binary,pool]==3.2.9",
  "python-dotenv==1.1.0",
  "dj-database-url==3.0.0",
  "django-extensions==4.1",
  "requests==2.32.3",
  "quart==0.20.0",
  "quart-db==0.9.0",
  "hypercorn==0.17.3",
  "fudgeo==1.2.0",
  "azure-storage-blob==12.25.1",
]

[dependency-groups]
dev = [
  "ipython>=8.31.0",
  "ipdb>=0.13.13",
  "mixer>=7.2.2",
  "pre-commit>=4.1.0",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.0.0",
]

# Reference: https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
ignore = [
  "E501", # Line too long
  "E722", # Bare except
]

# Reference: https://www.djlint.com/docs/configuration/
[tool.djlint]
profile = "django"

[tool.pytest.ini_options]
testpaths = ["tests"]
